this is important -- actually making good tests is a different lesson.
I sure like kotlins. is good. is very good boy.

i also like dogs.

+++ M03/guided/app/src/main/java/com/saucefan/stuff/gcalc/Math.kt
@@ -3,13 +3,16 @@ package com.saucefan.stuff.gcalc
 class Math {
 
     fun addIntegers(a:Int,b:Int):Int {
-        //todo:Write Implementation
+
         return a+b
     }
 
-    fun multiplyIntegers(x:Int,y:Int):Int {
+    fun multiplyIntegers(x:Long,y:Long):Long {
         return x*y
 
     }
+
+
+
 
 }
